+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Unreleased changes
+
+
+- The sdk now requires a `rustc` version at least 1.67 (Before it required version 1.66).
+- Add a `contract_update` helper analogous to `contract_init` to extract an
+  execution tree from a smart contract update transaction.
+- Add a `ccd_cost` helper to `ChainParameters` to convert NRG cost to CCD.
 - Add support for `DryRun`. Requires a node version at least 6.2.
+
+## 3.1.0
+
 - Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates
   of the baker for the reward period. Requires a node version at least 6.1.
 - Add support for `GetWinningBakersEpoch`. Requires a node version at least 6.1.
@@ -7,7 +16,6 @@
 - Add support for `GetBakersRewardPeriod` endpoint. Requires a node version at least 6.1.
 - Add Support for `GetBakerEarliestWinTime` endpoint. Requires a node version at least 6.1.
 - Add support for `GetBlockCertificates`. Requires a node version at least 6.1.
-- Update minimum supported rust version to `1.66`.
 - Add `make_update` and `make_update_raw` methods to the `ContractClient`. They
   are like `update` and `update_raw` but instead of sending a transaction they
   only construct it and return it.
@@ -18,6 +26,8 @@
 - Add `make_transfer` and `make_update_operator` functions to the CIS2 client.
   These are like the methods without the `make_`, except that they only
   construct the transaction.
+- Update minimum supported rust version to `1.66`.
+
 
 ## 3.0.1